backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
redis==5.0.5
celery==5.4.0
celery[redis]==5.4.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
flower==2.0.1
